added capacity limit in messoption

+++ swd/main/views.py
@@ -457,20 +457,30 @@ def messoption(request):
                     errors.append(obj)
             context['errors'] = errors
 
-        if (vacations.count() and len(errors) == 0) or (vacations.count() == 0) or (edit):
+        if (vacations.count() and len(errors) == 0) or (vacations.count() == 0) or (edit and len(errors) == 0):
             # Mess Option Filling
             mess = request.POST.get('mess')
             if edit:
                 messoption.delete()
-            messoptionfill = MessOption(
-                student=student,
-                monthYear=messopen[0].monthYear,
-                mess=mess)
-            messoptionfill.save()
-
-        if created or (vacations.count() == 0):
-            return redirect('messoption')
-
+            if messopen[0].capacity != None:
+                if MessOption.objects.filter(mess=mess).count() < messopen[0].capacity:
+                    messoptionfill = MessOption(
+                        student=student,
+                        monthYear=messopen[0].monthYear,
+                        mess=mess)
+                    messoptionfill.save()
+                    context['mess'] = messoption[0]
+                    context['option'] = 1
+                else:
+                    context['capacity']="Choose different mess, capacity full."
+            else:
+                messoptionfill = MessOption(
+                        student=student,
+                        monthYear=messopen[0].monthYear,
+                        mess=mess)
+                messoptionfill.save()
+        # if created or (vacations.count() == 0):
+        #     return redirect('messoption')
     return render(request, "mess.html", context)
